+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This file is used to list changes made in each version of the aws-parallelcluste
 
 **ENHANCEMENTS**
 - Add support for launching nodes across multiple availability zones to increase capacity availability.
+- Add logging of compute node console output to CloudWatch from head node on compute node bootstrap failure.
+
 
 **CHANGES**
 - Do not consider dynamic nodes in IDLE+CLOUD+COMPLETING+POWER_DOWN+NOT_RESPONDING as unhealthy anymore.